+++ NEWS.md
@@ -39,6 +39,7 @@
   bugs (#1178).
 - Added support for conditional force assignment (#1185).
 - Access and file types are now supported in VHPI (#1192).
+- Fixed a crash when open actual was used for formal parameter without default.
 - Several other minor bugs were resolved (#1174, #1159, #1177, #1191).
 
 ## Version 1.15.2 - 2025-03-01